requirements.txt
# Production Dependencies for Literature Review System
# Install with: pip install -r requirements.txt

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
google-generativeai>=0.3.0
google-genai>=0.1.0
python-dotenv>=1.0.0

# PDF processing
pdfplumber>=0.10.0
pypdf>=3.15.0
PyMuPDF>=1.23.0  # Enhanced PDF metadata extraction

# HTTP requests (for DOI verification)
requests>=2.31.0

# HTML parsing
beautifulsoup4>=4.12.0

# Visualization
plotly>=5.14.0
matplotlib>=3.7.0

# NLP and embeddings
sentence-transformers>=2.2.0

# Graph analysis
networkx>=3.0

# Scientific computing
scipy>=1.11.0
scikit-learn>=1.3.0
